+++ src/WriterXml.cs
@@ -475,6 +475,29 @@ namespace Dec
             }
         }
 
+        public override void WriteQueue(IEnumerable value)
+        {
+            // We actually just treat this like an array right now; it's the same behavior and it's easier
+            Type keyType = value.GetType().GetGenericArguments()[0];
+            var array = value.GetType().GetMethod("ToArray").Invoke(value, new object[] { }) as Array;
+
+            WriteArray(array);
+        }
+
+        public override void WriteStack(IEnumerable value)
+        {
+            // We actually just treat this like an array right now; it's the same behavior and it's easier
+            Type keyType = value.GetType().GetGenericArguments()[0];
+            var array = value.GetType().GetMethod("ToArray").Invoke(value, new object[] { }) as Array;
+
+            // For some reason this writes it out to an array in the reverse order than I'd expect
+            // (and also the reverse order it inputs in!)
+            // so, uh, time to munge
+            Array.Reverse(array);
+
+            WriteArray(array);
+        }
+
         public override void WriteTuple(object value, System.Runtime.CompilerServices.TupleElementNamesAttribute names)
         {
             var args = value.GetType().GenericTypeArguments;